+++ modules/registrars/synergywholesaledomains/synergywholesaledomains.php
@@ -281,6 +281,33 @@ function synergywholesaledomains_getConfigArray(array $params)
             'Type' => 'yesno',
             'Description' => 'Enable this option to force update the WHOIS.json data<br><b>NOTE:</b> This option will be disabled automatically again once you have clicked \'Save Changes\' and the update sequence is completed.',
         ],
+        'defaultDnsConfig' => [
+            'FriendlyName' => 'Default DNS Config',
+            'Type' => 'dropdown',
+            'Options' => [
+                '0' => 'Nothing',
+                '1' => 'Nameservers',
+                '2' => 'FreeDNS with email forwarding',
+                '3' => 'Parked',
+                '4' => 'FreeDNS',
+                '5' => 'SWS Account Default',
+                '6' => 'Legacy Hosting',
+                '7' => 'Wholesale Hosting'
+            ],
+            'Description' => 'Which Default DNS Config will be applied to newly registered domains',
+        ],
+        'enableDnsManagement' => [
+            'FriendlyName' => 'Enable DNS Management',
+            'Type' => 'yesno',
+            'Size' => '1',
+            'Description' => 'Tick if you wish to enable DNS management on the domain, if Default DNS supports it.',
+        ],
+        'enableEmailForwarding' => [
+            'FriendlyName' => 'Enable Email Forwarding',
+            'Type' => 'yesno',
+            'Size' => '1',
+            'Description' => 'Tick if you wish to enable email forwarding on the domain, if Default DNS supports it.',
+        ],
         'Version' => [
             'Description' => 'This module version: ' . SW_MODULE_VERSION,
         ],
@@ -611,9 +638,10 @@ function synergywholesaledomains_RegisterDomain(array $params)
         $request['costPrice'] = $params['premiumCost'];
         $request['premium'] = true;
     }
-    
+
     try {
         synergywholesaledomains_apiRequest('domainRegister', $params, $request);
+
         return [
             'success' => true,
         ];
@@ -938,6 +966,18 @@ function synergywholesaledomains_Sync(array $params)
                     $returnData = [
                         'transferredAway' => true,
                     ];
+                    break;
+                case 'register_manual':
+                    $returnData = [
+                        'active' => false,
+                        'cancelled' => false,
+                        'transferredAway' => false,
+                    ];
+                    Capsule::table('tbldomains')
+                        ->where('id', $params['domainid'])
+                        ->update([
+                            'status' => 'Pending Registration',
+                        ]);
                     break;
                 default:
                     $returnData = [
@@ -2501,7 +2541,7 @@ function synergywholesaledomains_adhocTransferSync(array $params, $domainInfo)
         'message' => nl2br(
             empty($syncMessages) ?
             'Domain Sync successful.' :
-            'Updated;\n    - ' . implode('\n    - ', $syncMessages)
+            "Updated;\n    - " . implode("\n    - ", $syncMessages)
         )
     ];
 }
@@ -2782,15 +2822,21 @@ if (class_exists('\WHMCS\Domain\TopLevel\ImportItem') && class_exists('\WHMCS\Re
         foreach ($response['pricing'] as $extension) {
             $tld = '.' . $extension->tld;
             $transfer_price = $extension->transfer;
+            $register_price = $extension->register_1_year;
+
             if (preg_match('/\.au$/', $tld)) {
                 $transfer_price = 0.00;
+            }
+
+            if ($register_price < $extension->renew) {
+                $register_price = $extension->renew;
             }
 
             $results[] = (new WHMCS\Domain\TopLevel\ImportItem())
                 ->setExtension($tld)
                 ->setMinYears($extension->minPeriod)
                 ->setMaxYears($extension->maxPeriod)
-                ->setRegisterPrice($extension->register_1_year)
+                ->setRegisterPrice($register_price)
                 ->setRenewPrice($extension->renew)
                 ->setTransferPrice($transfer_price)
                 ->setRedemptionFeePrice($extension->redemption)